[DDW-423] Removing ADA store extends
